+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/combobox/FocusableComboBox.java
@@ -43,6 +43,16 @@ final class FocusableComboBox<T> extends JComboBox<T> {
 	}
 
 	@Override
+	public boolean requestFocusInWindow() {
+		if (isEditable()) {
+			return getEditor().getEditorComponent().requestFocusInWindow();
+		}
+		else {
+			return super.requestFocusInWindow();
+		}
+	}
+
+	@Override
 	public synchronized void addFocusListener(FocusListener listener) {
 		super.addFocusListener(listener);
 		if (isEditable()) {