+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/spinner/AbstractSpinnerBuilder.java
@@ -160,5 +160,10 @@ abstract class AbstractSpinnerBuilder<T, B extends SpinnerBuilder<T, B>> extends
 		public void requestFocus() {
 			getEditor().requestFocus();
 		}
+
+		@Override
+		public boolean requestFocusInWindow() {
+			return getEditor().requestFocusInWindow();
+		}
 	}
 }