+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -351,6 +351,7 @@ public class EntityPanel extends JPanel {
 						.nonNull(configuration.initialEditState)
 						.consumer(this::updateEditState)
 						.build();
+		displayWhenActivated();
 		createControls();
 	}
 
@@ -1139,6 +1140,22 @@ public class EntityPanel extends JPanel {
 		controlMap.control(NAVIGATE_RIGHT).set(command(new Navigate(RIGHT)));
 		if (containsTablePanel()) {
 			controlMap.control(REFRESH).set(createRefreshTableControl());
+		}
+	}
+
+	private void displayWhenActivated() {
+		if (editPanel != null) {
+			editPanel.active().addConsumer(new DisplayActivatedPanel());
+		}
+	}
+
+	private final class DisplayActivatedPanel implements Consumer<Boolean> {
+
+		@Override
+		public void accept(Boolean active) {
+			if (active) {
+				display.request();
+			}
 		}
 	}
 
@@ -1953,7 +1970,9 @@ public class EntityPanel extends JPanel {
 
 		@Override
 		public void mouseClicked(MouseEvent e) {
-			editPanel.focus().afterUpdate().request();
+			if (!editPanel.active().is()) {
+				editPanel.focus().initial().request();
+			}
 		}
 	}